+++ src/html/mod.rs
@@ -165,7 +165,7 @@ impl Report for Html {
 
         try_else_return!(fs::mkdirp(&format!(
             "{}/{}/report/",
-            report_context.output_directory, id
+            report_context.output_directory, id.as_directory_name()
         )));
 
         let slope_estimate = &measurements.absolute_estimates[&Statistic::Slope];
@@ -244,7 +244,7 @@ impl Report for Html {
             &text,
             &format!(
                 "{}/{}/report/index.html",
-                report_context.output_directory, id
+                report_context.output_directory, id.as_directory_name()
             ),
         ));
     }
@@ -416,7 +416,7 @@ impl Html {
             data,
             measurements.avg_times,
             id,
-            format!("{}/{}/report/pdf.svg", context.output_directory, id),
+            format!("{}/{}/report/pdf.svg", context.output_directory, id.as_directory_name()),
             None,
         ));
         gnuplots.extend(plot::abs_distributions(
@@ -430,13 +430,13 @@ impl Html {
             point,
             (lb_, ub_),
             id,
-            format!("{}/{}/report/regression.svg", context.output_directory, id),
+            format!("{}/{}/report/regression.svg", context.output_directory, id.as_directory_name()),
             None,
             false,
         ));
         gnuplots.push(plot::pdf_small(
             &*measurements.avg_times,
-            format!("{}/{}/report/pdf_small.svg", context.output_directory, id),
+            format!("{}/{}/report/pdf_small.svg", context.output_directory, id.as_directory_name()),
             Some(THUMBNAIL_SIZE),
         ));
         gnuplots.push(plot::regression(
@@ -446,7 +446,7 @@ impl Html {
             id,
             format!(
                 "{}/{}/report/regression_small.svg",
-                context.output_directory, id
+                context.output_directory, id.as_directory_name()
             ),
             Some(THUMBNAIL_SIZE),
             true,
@@ -455,14 +455,14 @@ impl Html {
         if let Some(ref comp) = measurements.comparison {
             try_else_return!(fs::mkdirp(&format!(
                 "{}/{}/report/change/",
-                context.output_directory, id
+                context.output_directory, id.as_directory_name()
             )));
 
             let base_data = Data::new(&comp.base_iter_counts, &comp.base_sample_times);
 
             try_else_return!(fs::mkdirp(&format!(
                 "{}/{}/report/both",
-                context.output_directory, id
+                context.output_directory, id.as_directory_name()
             )));
             gnuplots.push(plot::both::regression(
                 base_data,
@@ -472,7 +472,7 @@ impl Html {
                 id,
                 format!(
                     "{}/{}/report/both/regression.svg",
-                    context.output_directory, id
+                    context.output_directory, id.as_directory_name()
                 ),
                 None,
                 false,
@@ -481,7 +481,7 @@ impl Html {
                 Sample::new(&comp.base_avg_times),
                 &*measurements.avg_times,
                 id,
-                format!("{}/{}/report/both/pdf.svg", context.output_directory, id),
+                format!("{}/{}/report/both/pdf.svg", context.output_directory, id.as_directory_name()),
                 None,
                 false,
             ));
@@ -506,7 +506,7 @@ impl Html {
                 id,
                 format!(
                     "{}/{}/report/relative_regression_small.svg",
-                    context.output_directory, id
+                    context.output_directory, id.as_directory_name()
                 ),
                 Some(THUMBNAIL_SIZE),
                 true,
@@ -517,7 +517,7 @@ impl Html {
                 id,
                 format!(
                     "{}/{}/report/relative_pdf_small.svg",
-                    context.output_directory, id
+                    context.output_directory, id.as_directory_name()
                 ),
                 Some(THUMBNAIL_SIZE),
                 true,